pyproject.toml
[tool.poetry]
name = "ttool"
version = "0.2.0"
description = "Tugys utils tools"
authors = ["IamTugy <tugmica@gmail.com>"]
readme = "README.md"
license = "MIT"

[project]
name = "ttool"
description = "Tugys utils tools"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
jira = "^3.4.0"
PyYAML = "^6.0"
dacite = "^1.6.0"
inquirer = "^2.9.2"
docopt = "^0.6.2"
typer = "^0.6.1"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"

[tool.poetry.scripts]
ttool = "ttool.main:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
